pyproject.toml
[tool.poetry]
name = "job-search"
version = "0.1.0"
description = "Получение данных о компаниях и вакансиях с сайта hh.ru"
authors = ["Maxim Ipatov <ipatovmaksim.1978@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
requests = "^2.32.3"
psycopg2 = "^2.9.9"
psycopg2-binary = "^2.9.9"
python-dotenv = "^1.0.1"
Pillow = "^10.4.0"
tqdm = "^4.66.5"
tabulate = "^0.9.0"
ipython = "^8.26.0"

[tool.poetry.group.dev.dependencies]
flake8 = "^7.1.1"
isort = "^5.13.2"
black = "^24.8.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

max-complexity = 10
per-file-ignores = { "__init__.py" = "F401" }

[tool.flake8]
line-length = 79
extend-ignore = ["E203"]
exclude = [
    "/__pycache__/",
    "env/",
    "venv/",
    ".env/",
    ".venv/",
    "tests",
    ".idea",
    ".vscode",
    ".git",
]

[tool.black]
line-length = 79
include = '\.pyi?$'


[tool.isort]
profile = "black"
line_length = 79
force_grid_wrap = 0
multi_line_output = 3
use_parentheses = true
include_trailing_comma = true
